v2.x (#54)
* Working on v2

* working on refactoring primitives and implementing self-validation logic

* streamlining type creation, list classes now extend their primitive parents, stubbing out validation stuff

* skipping overloaded properties during unserialization and serialization

* fixing serialization of -list types

+++ src/Enum/PrimitiveTypeEnum.php
@@ -62,7 +62,7 @@ class PrimitiveTypeEnum extends AbstractEnum
                 return $v;
 
             case PrimitiveTypeEnum::DECIMAL:
-                return 'float';
+                return 'double';
 
             case PrimitiveTypeEnum::POSITIVE_INTEGER:
             case PrimitiveTypeEnum::NEGATIVE_INTEGER:
